+++ frontend/src/pages/ProfilePage/profilePage.scss
@@ -1,44 +1,44 @@
 $mobile: 744px;
 $tablet: 1023px;
 $desktop: 1024px;
-
-label {
-  b {
-    color: var(--dark);
-    font-weight: 700;
-    font-size: 24px;
-    line-height: 28.13px;
-  }
-}
-
-.profile-page-wrapper {
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  user-select: none;
-}
+$navigation: 80px;
 
 .container {
   background-color: var(--white);
-  margin-top: 150px;
-  padding: 50px;
   width: 100%;
-  height: 100vh;
-  border-radius: 20px;
+  min-height: 100vh;
   display: flex;
   flex-direction: column;
   align-items: center;
+  justify-content: center;
+  user-select: none;
+  margin-top: $navigation;
+
+  label {
+    b {
+      color: var(--dark);
+      font-weight: 700;
+      font-size: 24px;
+      line-height: 28.13px;
+    }
+  }
 }
 
 .form {
   display: flex;
 }
 
-body form div {
+// TODO: class 충돌
+#root {
+  padding-bottom: 0;
+}
+body main .form {
   margin: 0;
 }
+body main .form div {
+  margin: 0;
+}
+// TODO: class 충돌
 
 .form .userImg {
   align-self: center;
@@ -58,6 +58,7 @@ body form div {
 
 .form > div > label {
   width: fit-content;
+  margin: 10px 0;
 }
 
 .field {
@@ -120,9 +121,10 @@ input[type='text']:focus {
 
 .stack-select {
   display: grid;
-  row-gap: 10px;
-  column-gap: 20px;
-  margin-top: 10px;
+  gap: 10px;
+  column-gap: 30px;
+  margin: 15px 0;
+
   grid-template-columns: repeat(4, 1fr);
 
   @media (max-width: $tablet) {
@@ -138,9 +140,9 @@ input[type='text']:focus {
     cursor: pointer;
   }
 
-  input {
-    margin-right: 8px;
-    border-radius: 15px;
+  div {
+    display: flex;
+    justify-content: space-between;
   }
 }
 
@@ -151,7 +153,7 @@ input[type='text']:focus {
 .clearfix {
   display: flex;
   justify-content: flex-end;
-  margin-top: 50px;
+
 
   button {
     border-radius: 10px;
