+++ frontend/src/router/Router.js
@@ -52,6 +52,11 @@ class Router {
       RouterContext.setState({ params });
       const Page = this.routes[i].component;
       window.scrollTo(0, 0);
+
+      if (currentPath[0] === 'profile') {
+        new Page({ container: this.target });
+        return;
+      }
       new Page(this.target);
       return;
     }
